Revert "Merge branch 'main' into splash/signon"
This reverts commit ba8fcf204496e1b65a1ee5ba9a1b7ff3498bb8d0, reversing
changes made to 32acd8a803fde9267afe1746ce7fe02f4b5fa7d5.
