+++ README.md
@@ -1,41 +1,19 @@
-# Evil Task Manager
+# Task Manager
 
 ## Overview
 
-This is an Evil Task Manager to help delegate evil jobs and assignments. To make the world a better place in our vision. The correct vision.
+This is a task manager to help delegate jobs and assignments
 
 ## Features
 
-- Add projects
-- Add, edit tasks within projects
+- Add, edit, and delete tasks
 - Mark tasks as completed
+- Filter tasks based on status (completed, pending)
 - Responsive design for seamless use on various devices
 
 ## Technologies Used
 
-### Frontend
-
-- React
-- HTML/CSS
-- JavaScript
-
-### Backend
-
-- Node.js
-- MongoDB
-- Express
-
-## Resources
-
-- Task Management MERN Project
-- Backend server set up - https://www.youtube.com/watch?v=RHLxtAo0aEI&t=127s
-- Mongo DB documentation - https://www.mongodb.com/docs/atlas/getting-started/
-- Vercel Documentation - https://vercel.com/docs
-- React documentation - https://legacy.reactjs.org/docs/getting-started.html
-
-## Contributors
-
-- [Oscar Romero](https://github.com/oromero7153)
-- [Jose Rodrigues](https://github.com/jorodrig96)
-- [Manny Delgado](https://github.com/MannyGDP)
-- [Brandon Iida](https://github.com/oallelsefailo)
+- **Frontend:** React
+- **Backend:** Node.js with Express
+- **Database:** MongoDB
+- **ODM:** Mongoose
