+++ client/src/App.js
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from "react";
 import "./App.css";
 import Navigation from "./components/Navigation";
 import Container from "./components/Container";
-// import LoginPage from "./components/LoginPage";
+// import LoginPage from "./components/LoginPage"; 
+import Submitter from "./components/Submit";
 import PacmanLoader from "react-spinners/PacmanLoader";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import LoginPage from "./components/LoginPage";
@@ -41,7 +42,11 @@ function App() {
               {isLoggedIn && <Navigation />} 
               <Routes>
                 <Route path="/task" element={<Container />} />
-                {/* <Route path="/" element={<LoginPage onLogin={handleLogin} />} /> */}
+                <Route path="/task/submit" element={<Submitter />} />
+                <Route path="/" element={<SignupLogIn/>} />
+                <Route path="/login" element={<LoginPage onLogin={handleLogin} />} /> 
+                <Route path="/signup" element={<SignupPage/>} />
+                {/*<Route path="/" element={<LoginPage onLogin={handleLogin} />} /> */}
               </Routes>
 
             </div>
