+++ client/src/components/container.css
@@ -1,11 +1,9 @@
 .container {
-  flex-shrink: 0;
-  flex-basis: auto;
+  flex-grow: 1;
   padding: 10px;
-  border: 1px solid var(--text-color);
+  border: 1px solid #ddd;
   border-radius: 8px;
-  width: calc(100vw - 170px);
-  min-height: 200px;
+  width: calc(100vw - 120px);
   box-sizing: border-box;
 }
 
@@ -15,7 +13,7 @@
 }
 
 .title img {
-  height: 45px; 
+  height: 60px; 
   margin-right: 10px; 
 }
 
@@ -44,6 +42,7 @@
   border: none;
   border-radius: 8px;
   cursor: pointer;
+  display: block;
 }
 
 a {
@@ -54,97 +53,21 @@ a {
   background-color: #394046; 
 }
 
-/* Task Board */
-
-.task {
-  display: flex;
-  flex-wrap: wrap; 
-  gap: 20px; 
-  justify-content: flex-start;
-  align-items: stretch; 
-}
-
-.task li {
-  width: calc(20% - 20px);
-  display: block; 
-  margin-bottom: 20px; 
-}
-
-.task-content {
-  position: relative;
-  display: flex; 
-  align-items: center;
-  width: 100%;
-  justify-content: start;
-}
-
-.task ul {
-  width: 100%;
-  padding: 0;
-  list-style: none;
-  display: flex; 
-  flex-wrap: wrap; 
-  justify-content: flex-start;
-  gap: 20px;
-}
-
-.task img {
-  width: 100%; 
-  height: 160px;
-  object-fit: cover;
-  margin-right: 0; 
-}
-
-.task p {
-  position: absolute; 
-  top: 70%; 
-  left: 50%; 
-  transform: translate(-50%, -50%); 
-  color: white; 
-  background-color: rgba(0, 0, 0, 0.5); 
-  padding: 5px; 
-  width: calc(100% - 9px); 
-  text-align: center; 
-  white-space: normal; 
-  overflow: hidden; 
-}
-
-/* Keyframes */
-
-@keyframes evil-red {
-  0%, 100% { 
-    color: #FFFFFF; /* White */
-    text-shadow: 0 0 8px #FFFFFF; /* White Glow */
-  } 
-  25% { 
-    color: #FFCCCC; /* Very Light Red */
-    text-shadow: 0 0 10px #FFCCCC;
-  } 
-  50% { 
-    color: #990000; /* Dark Red */
-    text-shadow: 0 0 18px #990000;
-  } 
-  75% { 
-    color: #FFCCCC; /* Very Light Red */
-    text-shadow: 0 0 10px #FFCCCC;
-  }
-}
-
-.task-content:hover p {
-  animation: evil-red 2s linear infinite;
-}
-
 /* Media Query */
 
 @media screen and (max-width: 768px) {
   .container {
-    width: calc(100vw - 50px);
+    width: calc(100vw - 30px);
     margin: 0 15px;
   }
 
   .title {
     flex-direction: row; 
     align-items: center; 
+  }
+
+  .title img {
+    height: 45px; 
   }
 
   .buttons {
@@ -157,43 +80,5 @@ a {
     margin-right: 0; 
     margin-bottom: 5px;
     padding: 5px;
-  }
-
-  .task li {
-    max-width: 100%; 
-  }
-
-  .task img {
-    width: 100%;
-    min-height: 150px;
-    max-height: 300px;
-  }
-
-  .task p {
-    font-size: 14px;
-    padding: 5px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-}
-
-@media screen and (max-width: 595px) {
-  .task li {
-    flex-basis: calc(50% - 10px); 
-    max-width: calc(50% - 10px); 
-  }
-  .title {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .buttons {
-    justify-content: flex-start;
-    margin-top: 10px; 
-  }
-
-  .task p {
-    font-size: 16px;
   }
 }
