--- client/src/components/modal.css
@@ -1,38 +1,0 @@
-.modal-container {
-    position: fixed; 
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5); 
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-}
-
-.modal-content {
-    position: relative;
-    background-color: #2c3034;
-    padding: 20px;
-    border-radius: 5px;
-    text-align: center;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    max-width: 500px; 
-    width: 100%; 
-    margin: 20px; 
-}
-
-.modal-content h1,
-.modal-content label,
-.modal-content input {
-    margin: 10px 0;
-}
-
-.close-button {
-    position: absolute;
-    top: -13px;
-    right: 10px;
-    cursor: pointer;
-    font-size: 20px;
-}
