+++ client/src/components/navigation.css
@@ -1,13 +1,9 @@
 header {
   display: flex;
   flex-direction: column;
-  justify-content: center; 
-  align-items: flex-start;
+  align-items: flex-start; 
   padding-left: 15px;
-  height: 100vh;
-  position: sticky;
-  top: 0; 
-  margin-right: 20px;
+  width: 15%;
 }
 
 .logo img {
